+++ kanidmd/lib/src/server/mod.rs
@@ -10,8 +10,8 @@ use std::time::Duration;
 
 use concread::arcache::{ARCache, ARCacheBuilder, ARCacheReadTxn};
 use concread::cowcell::*;
-use hashbrown::{HashMap, HashSet};
-use kanidm_proto::v1::{ConsistencyError, SchemaError, UiHint};
+use hashbrown::HashSet;
+use kanidm_proto::v1::{ConsistencyError, UiHint};
 use tokio::sync::{Semaphore, SemaphorePermit};
 use tracing::trace;
 
@@ -23,11 +23,11 @@ use crate::access::{
 use crate::be::{Backend, BackendReadTransaction, BackendTransaction, BackendWriteTransaction};
 // We use so many, we just import them all ...
 use crate::event::{
-    CreateEvent, DeleteEvent, ExistsEvent, ModifyEvent, ReviveRecycledEvent, SearchEvent,
+    CreateEvent, DeleteEvent, ExistsEvent, ModifyEvent, SearchEvent,
 };
 use crate::filter::{Filter, FilterInvalid, FilterValid, FilterValidResolved};
 use crate::identity::IdentityId;
-use crate::modify::{Modify, ModifyInvalid, ModifyList, ModifyValid};
+use crate::modify::{ModifyInvalid, ModifyList, ModifyValid};
 use crate::plugins::dyngroup::{DynGroup, DynGroupCache};
 use crate::plugins::Plugins;
 use crate::prelude::*;
@@ -38,12 +38,12 @@ use crate::schema::{
 };
 use crate::valueset::uuid_to_proto_string;
 
-use crate::server::modify::ModifyPartial;
-
 pub mod batch_modify;
 pub mod create;
 pub mod delete;
+pub mod migrations;
 pub mod modify;
+pub mod repl;
 pub mod search;
 
 const RESOLVE_FILTER_CACHE_MAX: usize = 4096;
@@ -1016,119 +1016,6 @@ impl QueryServer {
         }
     }
 
-    #[instrument(level = "info", name = "system_initialisation", skip_all)]
-    pub async fn initialise_helper(&self, ts: Duration) -> Result<(), OperationError> {
-        // Check our database version - attempt to do an initial indexing
-        // based on the in memory configuration
-        //
-        // If we ever change the core in memory schema, or the schema that we ship
-        // in fixtures, we have to bump these values. This is how we manage the
-        // first-run and upgrade reindexings.
-        //
-        // A major reason here to split to multiple transactions is to allow schema
-        // reloading to occur, which causes the idxmeta to update, and allows validation
-        // of the schema in the subsequent steps as we proceed.
-        let mut reindex_write_1 = self.write(ts).await;
-        reindex_write_1
-            .upgrade_reindex(SYSTEM_INDEX_VERSION)
-            .and_then(|_| reindex_write_1.commit())?;
-
-        // Because we init the schema here, and commit, this reloads meaning
-        // that the on-disk index meta has been loaded, so our subsequent
-        // migrations will be correctly indexed.
-        //
-        // Remember, that this would normally mean that it's possible for schema
-        // to be mis-indexed (IE we index the new schemas here before we read
-        // the schema to tell us what's indexed), but because we have the in
-        // mem schema that defines how schema is structuded, and this is all
-        // marked "system", then we won't have an issue here.
-        let mut ts_write_1 = self.write(ts).await;
-        ts_write_1
-            .initialise_schema_core()
-            .and_then(|_| ts_write_1.commit())?;
-
-        let mut ts_write_2 = self.write(ts).await;
-        ts_write_2
-            .initialise_schema_idm()
-            .and_then(|_| ts_write_2.commit())?;
-
-        // reindex and set to version + 1, this way when we bump the version
-        // we are essetially pushing this version id back up to step write_1
-        let mut reindex_write_2 = self.write(ts).await;
-        reindex_write_2
-            .upgrade_reindex(SYSTEM_INDEX_VERSION + 1)
-            .and_then(|_| reindex_write_2.commit())?;
-
-        // Force the schema to reload - this is so that any changes to index slope
-        // analysis are now reflected correctly.
-        //
-        // A side effect of these reloads is that other plugins or elements that reload
-        // on schema change are now setup.
-        let mut slope_reload = self.write(ts).await;
-        slope_reload.set_phase(ServerPhase::SchemaReady);
-        slope_reload.force_schema_reload();
-        slope_reload.commit()?;
-
-        // Now, based on the system version apply migrations. You may ask "should you not
-        // be doing migrations before indexes?". And this is a very good question! The issue
-        // is within a migration we must be able to search for content by pres index, and those
-        // rely on us being indexed! It *is* safe to index content even if the
-        // migration would cause a value type change (ie name changing from iutf8s to iname) because
-        // the indexing subsystem is schema/value agnostic - the fact the values still let their keys
-        // be extracted, means that the pres indexes will be valid even though the entries are pending
-        // migration. We must be sure to NOT use EQ/SUB indexes in the migration code however!
-        let mut migrate_txn = self.write(ts).await;
-        // If we are "in the process of being setup" this is 0, and the migrations will have no
-        // effect as ... there is nothing to migrate! It allows reset of the version to 0 to force
-        // db migrations to take place.
-        let system_info_version = match migrate_txn.internal_search_uuid(UUID_SYSTEM_INFO) {
-            Ok(e) => Ok(e.get_ava_single_uint32("version").unwrap_or(0)),
-            Err(OperationError::NoMatchingEntries) => Ok(0),
-            Err(r) => Err(r),
-        }?;
-        admin_debug!(?system_info_version);
-
-        if system_info_version < 3 {
-            migrate_txn.migrate_2_to_3()?;
-        }
-
-        if system_info_version < 4 {
-            migrate_txn.migrate_3_to_4()?;
-        }
-
-        if system_info_version < 5 {
-            migrate_txn.migrate_4_to_5()?;
-        }
-
-        if system_info_version < 6 {
-            migrate_txn.migrate_5_to_6()?;
-        }
-
-        if system_info_version < 7 {
-            migrate_txn.migrate_6_to_7()?;
-        }
-
-        if system_info_version < 8 {
-            migrate_txn.migrate_7_to_8()?;
-        }
-
-        if system_info_version < 9 {
-            migrate_txn.migrate_8_to_9()?;
-        }
-
-        migrate_txn.commit()?;
-        // Migrations complete. Init idm will now set the version as needed.
-
-        let mut ts_write_3 = self.write(ts).await;
-        ts_write_3.initialise_idm().and_then(|_| {
-            ts_write_3.set_phase(ServerPhase::Running);
-            ts_write_3.commit()
-        })?;
-        // TODO: work out if we've actually done any migrations before printing this
-        admin_debug!("Database version check and migrations success! ☀️  ");
-        Ok(())
-    }
-
     pub async fn verify(&self) -> Vec<Result<(), ConsistencyError>> {
         let mut r_txn = self.read().await;
         r_txn.verify()
@@ -1140,518 +1027,12 @@ impl<'a> QueryServerWriteTransaction<'a> {
         self.curtime
     }
 
-    #[instrument(level = "debug", skip_all)]
-    pub fn purge_tombstones(&self) -> Result<(), OperationError> {
-        // purge everything that is a tombstone.
-        let cid = self.cid.sub_secs(CHANGELOG_MAX_AGE).map_err(|e| {
-            admin_error!("Unable to generate search cid {:?}", e);
-            e
-        })?;
-
-        // Delete them - this is a TRUE delete, no going back now!
-        self.be_txn
-            .reap_tombstones(&cid)
-            .map_err(|e| {
-                admin_error!(err = ?e, "Tombstone purge operation failed (backend)");
-                e
-            })
-            .map(|_| {
-                admin_info!("Tombstone purge operation success");
-            })
-    }
-
-    #[instrument(level = "debug", skip_all)]
-    pub fn purge_recycled(&self) -> Result<(), OperationError> {
-        // Send everything that is recycled to tombstone
-        // Search all recycled
-        let cid = self.cid.sub_secs(RECYCLEBIN_MAX_AGE).map_err(|e| {
-            admin_error!(err = ?e, "Unable to generate search cid");
-            e
-        })?;
-        let rc = self.internal_search(filter_all!(f_and!([
-            f_eq("class", PVCLASS_RECYCLED.clone()),
-            f_lt("last_modified_cid", PartialValue::new_cid(cid)),
-        ])))?;
-
-        if rc.is_empty() {
-            admin_info!("No recycled present - purge operation success");
-            return Ok(());
-        }
-
-        // Modify them to strip all avas except uuid
-        let tombstone_cand: Result<Vec<_>, _> = rc
-            .iter()
-            .map(|e| {
-                e.to_tombstone(self.cid.clone())
-                    .validate(&self.schema)
-                    .map_err(|e| {
-                        admin_error!("Schema Violation in purge_recycled validate: {:?}", e);
-                        OperationError::SchemaViolation(e)
-                    })
-                    // seal if it worked.
-                    .map(|e| e.seal(&self.schema))
-            })
-            .collect();
-
-        let tombstone_cand = tombstone_cand?;
-
-        // Backend Modify
-        self.be_txn
-            .modify(&self.cid, &rc, &tombstone_cand)
-            .map_err(|e| {
-                admin_error!("Purge recycled operation failed (backend), {:?}", e);
-                e
-            })
-            .map(|_| {
-                admin_info!("Purge recycled operation success");
-            })
-    }
-
-    #[instrument(level = "debug", skip_all)]
-    pub fn revive_recycled(&mut self, re: &ReviveRecycledEvent) -> Result<(), OperationError> {
-        // Revive an entry to live. This is a specialised function, and draws a lot of
-        // inspiration from modify.
-        //
-        // Access is granted by the ability to ability to search the class=recycled
-        // and the ability modify + remove that class from the object.
-        if !re.ident.is_internal() {
-            security_info!(name = %re.ident, "revive initiator");
-        }
-
-        // Get the list of pre_candidates, using impersonate search.
-        let pre_candidates =
-            self.impersonate_search_valid(re.filter.clone(), re.filter.clone(), &re.ident)?;
-
-        // Is the list empty?
-        if pre_candidates.is_empty() {
-            if re.ident.is_internal() {
-                trace!(
-                    "revive: no candidates match filter ... continuing {:?}",
-                    re.filter
-                );
-                return Ok(());
-            } else {
-                request_error!(
-                    "revive: no candidates match filter, failure {:?}",
-                    re.filter
-                );
-                return Err(OperationError::NoMatchingEntries);
-            }
-        };
-
-        trace!("revive: pre_candidates -> {:?}", pre_candidates);
-
-        // Check access against a "fake" modify.
-        let modlist = ModifyList::new_list(vec![Modify::Removed(
-            AttrString::from("class"),
-            PVCLASS_RECYCLED.clone(),
-        )]);
-
-        let m_valid = modlist.validate(self.get_schema()).map_err(|e| {
-            admin_error!("revive recycled modlist Schema Violation {:?}", e);
-            OperationError::SchemaViolation(e)
-        })?;
-
-        let me =
-            ModifyEvent::new_impersonate(&re.ident, re.filter.clone(), re.filter.clone(), m_valid);
-
-        let access = self.get_accesscontrols();
-        let op_allow = access
-            .modify_allow_operation(&me, &pre_candidates)
-            .map_err(|e| {
-                admin_error!("Unable to check modify access {:?}", e);
-                e
-            })?;
-        if !op_allow {
-            return Err(OperationError::AccessDenied);
-        }
-
-        // Are all of the entries actually recycled?
-        if pre_candidates.iter().all(|e| e.mask_recycled().is_some()) {
-            admin_warn!("Refusing to revive entries that are already live!");
-            return Err(OperationError::AccessDenied);
-        }
-
-        // Build the list of mods from directmo, to revive memberships.
-        let mut dm_mods: HashMap<Uuid, ModifyList<ModifyInvalid>> =
-            HashMap::with_capacity(pre_candidates.len());
-
-        for e in &pre_candidates {
-            // Get this entries uuid.
-            let u: Uuid = e.get_uuid();
-
-            if let Some(riter) = e.get_ava_as_refuuid("directmemberof") {
-                for g_uuid in riter {
-                    dm_mods
-                        .entry(g_uuid)
-                        .and_modify(|mlist| {
-                            let m = Modify::Present(AttrString::from("member"), Value::Refer(u));
-                            mlist.push_mod(m);
-                        })
-                        .or_insert({
-                            let m = Modify::Present(AttrString::from("member"), Value::Refer(u));
-                            ModifyList::new_list(vec![m])
-                        });
-                }
-            }
-        }
-
-        // clone the writeable entries.
-        let mut candidates: Vec<Entry<EntryInvalid, EntryCommitted>> = pre_candidates
-            .iter()
-            .map(|er| er.as_ref().clone().invalidate(self.cid.clone()))
-            // Mutate to apply the revive.
-            .map(|er| er.to_revived())
-            .collect();
-
-        // Are they all revived?
-        if candidates.iter().all(|e| e.mask_recycled().is_none()) {
-            admin_error!("Not all candidates were correctly revived, unable to proceed");
-            return Err(OperationError::InvalidEntryState);
-        }
-
-        // Do we need to apply pre-mod?
-        // Very likely, incase domain has renamed etc.
-        Plugins::run_pre_modify(self, &mut candidates, &me).map_err(|e| {
-            admin_error!("Revive operation failed (plugin), {:?}", e);
-            e
-        })?;
-
-        // Schema validate
-        let res: Result<Vec<Entry<EntrySealed, EntryCommitted>>, OperationError> = candidates
-            .into_iter()
-            .map(|e| {
-                e.validate(&self.schema)
-                    .map_err(|e| {
-                        admin_error!("Schema Violation {:?}", e);
-                        OperationError::SchemaViolation(e)
-                    })
-                    .map(|e| e.seal(&self.schema))
-            })
-            .collect();
-
-        let norm_cand: Vec<Entry<_, _>> = res?;
-
-        // build the mod partial
-        let mp = ModifyPartial {
-            norm_cand,
-            pre_candidates,
-            me: &me,
-        };
-
-        // Call modify_apply
-        self.modify_apply(mp)?;
-
-        // If and only if that succeeds, apply the direct membership modifications
-        // if possible.
-        for (g, mods) in dm_mods {
-            // I think the filter/filter_all shouldn't matter here because the only
-            // valid direct memberships should be still valid/live references, as refint
-            // removes anything that was deleted even from recycled entries.
-            let f = filter_all!(f_eq("uuid", PartialValue::Uuid(g)));
-            self.internal_modify(&f, &mods)?;
-        }
-
-        Ok(())
-    }
-
-    #[instrument(level = "debug", skip_all)]
-    pub fn revive_recycled_legacy(
-        &mut self,
-        re: &ReviveRecycledEvent,
-    ) -> Result<(), OperationError> {
-        // Revive an entry to live. This is a specialised function, and draws a lot of
-        // inspiration from modify.
-        //
-        //
-        // Access is granted by the ability to ability to search the class=recycled
-        // and the ability modify + remove that class from the object.
-
-        // create the modify for access testing.
-        // tl;dr, remove the class=recycled
-        let modlist = ModifyList::new_list(vec![Modify::Removed(
-            AttrString::from("class"),
-            PVCLASS_RECYCLED.clone(),
-        )]);
-
-        let m_valid = modlist.validate(self.get_schema()).map_err(|e| {
-            admin_error!(
-                "Schema Violation in revive recycled modlist validate: {:?}",
-                e
-            );
-            OperationError::SchemaViolation(e)
-        })?;
-
-        // Get the entries we are about to revive.
-        //    we make a set of per-entry mod lists. A list of lists even ...
-        let revive_cands =
-            self.impersonate_search_valid(re.filter.clone(), re.filter.clone(), &re.ident)?;
-
-        let mut dm_mods: HashMap<Uuid, ModifyList<ModifyInvalid>> =
-            HashMap::with_capacity(revive_cands.len());
-
-        for e in revive_cands {
-            // Get this entries uuid.
-            let u: Uuid = e.get_uuid();
-
-            if let Some(riter) = e.get_ava_as_refuuid("directmemberof") {
-                for g_uuid in riter {
-                    dm_mods
-                        .entry(g_uuid)
-                        .and_modify(|mlist| {
-                            let m = Modify::Present(AttrString::from("member"), Value::Refer(u));
-                            mlist.push_mod(m);
-                        })
-                        .or_insert({
-                            let m = Modify::Present(AttrString::from("member"), Value::Refer(u));
-                            ModifyList::new_list(vec![m])
-                        });
-                }
-            }
-        }
-
-        // Now impersonate the modify
-        self.impersonate_modify_valid(re.filter.clone(), re.filter.clone(), m_valid, &re.ident)?;
-        // If and only if that succeeds, apply the direct membership modifications
-        // if possible.
-        for (g, mods) in dm_mods {
-            // I think the filter/filter_all shouldn't matter here because the only
-            // valid direct memberships should be still valid/live references.
-            let f = filter_all!(f_eq("uuid", PartialValue::Uuid(g)));
-            self.internal_modify(&f, &mods)?;
-        }
-        Ok(())
-    }
-
     #[allow(clippy::mut_from_ref)]
     pub(crate) fn get_dyngroup_cache(&self) -> &mut DynGroupCache {
         unsafe {
             let mptr = self.dyngroup_cache.as_ptr();
             (*mptr).get_mut()
         }
-    }
-
-    /// Migrate 2 to 3 changes the name, domain_name types from iutf8 to iname.
-    #[instrument(level = "debug", skip_all)]
-    pub fn migrate_2_to_3(&mut self) -> Result<(), OperationError> {
-        admin_warn!("starting 2 to 3 migration. THIS MAY TAKE A LONG TIME!");
-        // Get all entries where pres name or domain_name. INCLUDE TS + RECYCLE.
-
-        let filt = filter_all!(f_or!([f_pres("name"), f_pres("domain_name"),]));
-
-        let pre_candidates = self.internal_search(filt).map_err(|e| {
-            admin_error!(err = ?e, "migrate_2_to_3 internal search failure");
-            e
-        })?;
-
-        // If there is nothing, we donn't need to do anything.
-        if pre_candidates.is_empty() {
-            admin_info!("migrate_2_to_3 no entries to migrate, complete");
-            return Ok(());
-        }
-
-        // Change the value type.
-        let mut candidates: Vec<Entry<EntryInvalid, EntryCommitted>> = pre_candidates
-            .iter()
-            .map(|er| er.as_ref().clone().invalidate(self.cid.clone()))
-            .collect();
-
-        candidates.iter_mut().try_for_each(|er| {
-            let nvs = if let Some(vs) = er.get_ava_set("name") {
-                vs.migrate_iutf8_iname()?
-            } else {
-                None
-            };
-            if let Some(nvs) = nvs {
-                er.set_ava_set("name", nvs)
-            }
-
-            let nvs = if let Some(vs) = er.get_ava_set("domain_name") {
-                vs.migrate_iutf8_iname()?
-            } else {
-                None
-            };
-            if let Some(nvs) = nvs {
-                er.set_ava_set("domain_name", nvs)
-            }
-
-            Ok(())
-        })?;
-
-        // Schema check all.
-        let res: Result<Vec<Entry<EntrySealed, EntryCommitted>>, SchemaError> = candidates
-            .into_iter()
-            .map(|e| e.validate(&self.schema).map(|e| e.seal(&self.schema)))
-            .collect();
-
-        let norm_cand: Vec<Entry<_, _>> = match res {
-            Ok(v) => v,
-            Err(e) => {
-                admin_error!("migrate_2_to_3 schema error -> {:?}", e);
-                return Err(OperationError::SchemaViolation(e));
-            }
-        };
-
-        // Write them back.
-        self.be_txn
-            .modify(&self.cid, &pre_candidates, &norm_cand)
-            .map_err(|e| {
-                admin_error!("migrate_2_to_3 modification failure -> {:?}", e);
-                e
-            })
-        // Complete
-    }
-
-    /// Migrate 3 to 4 - this triggers a regen of the domains security token
-    /// as we previously did not have it in the entry.
-    #[instrument(level = "debug", skip_all)]
-    pub fn migrate_3_to_4(&mut self) -> Result<(), OperationError> {
-        admin_warn!("starting 3 to 4 migration.");
-        let filter = filter!(f_eq("uuid", (*PVUUID_DOMAIN_INFO).clone()));
-        let modlist = ModifyList::new_purge("domain_token_key");
-        self.internal_modify(&filter, &modlist)
-        // Complete
-    }
-
-    /// Migrate 4 to 5 - this triggers a regen of all oauth2 RS es256 der keys
-    /// as we previously did not generate them on entry creation.
-    #[instrument(level = "debug", skip_all)]
-    pub fn migrate_4_to_5(&mut self) -> Result<(), OperationError> {
-        admin_warn!("starting 4 to 5 migration.");
-        let filter = filter!(f_and!([
-            f_eq("class", (*PVCLASS_OAUTH2_RS).clone()),
-            f_andnot(f_pres("es256_private_key_der")),
-        ]));
-        let modlist = ModifyList::new_purge("es256_private_key_der");
-        self.internal_modify(&filter, &modlist)
-        // Complete
-    }
-
-    /// Migrate 5 to 6 - This updates the domain info item to reset the token
-    /// keys based on the new encryption types.
-    #[instrument(level = "debug", skip_all)]
-    pub fn migrate_5_to_6(&mut self) -> Result<(), OperationError> {
-        admin_warn!("starting 5 to 6 migration.");
-        let filter = filter!(f_eq("uuid", (*PVUUID_DOMAIN_INFO).clone()));
-        let mut modlist = ModifyList::new_purge("domain_token_key");
-        // We need to also push the version here so that we pass schema.
-        modlist.push_mod(Modify::Present(
-            AttrString::from("version"),
-            Value::Uint32(0),
-        ));
-        self.internal_modify(&filter, &modlist)
-        // Complete
-    }
-
-    /// Migrate 6 to 7
-    ///
-    /// Modify accounts that are not persons, to be service accounts so that the extension
-    /// rules remain valid.
-    #[instrument(level = "debug", skip_all)]
-    pub fn migrate_6_to_7(&mut self) -> Result<(), OperationError> {
-        admin_warn!("starting 6 to 7 migration.");
-        let filter = filter!(f_and!([
-            f_eq("class", (*PVCLASS_ACCOUNT).clone()),
-            f_andnot(f_eq("class", (*PVCLASS_PERSON).clone())),
-        ]));
-        let modlist = ModifyList::new_append("class", Value::new_class("service_account"));
-        self.internal_modify(&filter, &modlist)
-        // Complete
-    }
-
-    /// Migrate 7 to 8
-    ///
-    /// Touch all service accounts to trigger a regen of their es256 jws keys for api tokens
-    #[instrument(level = "debug", skip_all)]
-    pub fn migrate_7_to_8(&mut self) -> Result<(), OperationError> {
-        admin_warn!("starting 7 to 8 migration.");
-        let filter = filter!(f_eq("class", (*PVCLASS_SERVICE_ACCOUNT).clone()));
-        let modlist = ModifyList::new_append("class", Value::new_class("service_account"));
-        self.internal_modify(&filter, &modlist)
-        // Complete
-    }
-
-    /// Migrate 8 to 9
-    ///
-    /// This migration updates properties of oauth2 relying server properties. First, it changes
-    /// the former basic value to a secret utf8string.
-    ///
-    /// The second change improves the current scope system to remove the implicit scope type.
-    #[instrument(level = "debug", skip_all)]
-    pub fn migrate_8_to_9(&mut self) -> Result<(), OperationError> {
-        admin_warn!("starting 8 to 9 migration.");
-        let filt = filter_all!(f_or!([
-            f_eq("class", PVCLASS_OAUTH2_RS.clone()),
-            f_eq("class", PVCLASS_OAUTH2_BASIC.clone()),
-        ]));
-
-        let pre_candidates = self.internal_search(filt).map_err(|e| {
-            admin_error!(err = ?e, "migrate_8_to_9 internal search failure");
-            e
-        })?;
-
-        // If there is nothing, we donn't need to do anything.
-        if pre_candidates.is_empty() {
-            admin_info!("migrate_8_to_9 no entries to migrate, complete");
-            return Ok(());
-        }
-
-        // Change the value type.
-        let mut candidates: Vec<Entry<EntryInvalid, EntryCommitted>> = pre_candidates
-            .iter()
-            .map(|er| er.as_ref().clone().invalidate(self.cid.clone()))
-            .collect();
-
-        candidates.iter_mut().try_for_each(|er| {
-            // Migrate basic secrets if they exist.
-            let nvs = er
-                .get_ava_set("oauth2_rs_basic_secret")
-                .and_then(|vs| vs.as_utf8_iter())
-                .and_then(|vs_iter| {
-                    ValueSetSecret::from_iter(vs_iter.map(|s: &str| s.to_string()))
-                });
-            if let Some(nvs) = nvs {
-                er.set_ava_set("oauth2_rs_basic_secret", nvs)
-            }
-
-            // Migrate implicit scopes if they exist.
-            let nv = if let Some(vs) = er.get_ava_set("oauth2_rs_implicit_scopes") {
-                vs.as_oauthscope_set()
-                    .map(|v| Value::OauthScopeMap(UUID_IDM_ALL_PERSONS, v.clone()))
-            } else {
-                None
-            };
-
-            if let Some(nv) = nv {
-                er.add_ava("oauth2_rs_scope_map", nv)
-            }
-            er.purge_ava("oauth2_rs_implicit_scopes");
-
-            Ok(())
-        })?;
-
-        // Schema check all.
-        let res: Result<Vec<Entry<EntrySealed, EntryCommitted>>, SchemaError> = candidates
-            .into_iter()
-            .map(|e| e.validate(&self.schema).map(|e| e.seal(&self.schema)))
-            .collect();
-
-        let norm_cand: Vec<Entry<_, _>> = match res {
-            Ok(v) => v,
-            Err(e) => {
-                admin_error!("migrate_8_to_9 schema error -> {:?}", e);
-                return Err(OperationError::SchemaViolation(e));
-            }
-        };
-
-        // Write them back.
-        self.be_txn
-            .modify(&self.cid, &pre_candidates, &norm_cand)
-            .map_err(|e| {
-                admin_error!("migrate_8_to_9 modification failure -> {:?}", e);
-                e
-            })
-        // Complete
     }
 
     // These are where searches and other actions are actually implemented. This
@@ -3054,194 +2435,6 @@ mod tests {
     }
 
     #[qs_test]
-    async fn test_recycle_simple(server: &QueryServer) {
-        // First we setup some timestamps
-        let time_p1 = duration_from_epoch_now();
-        let time_p2 = time_p1 + Duration::from_secs(RECYCLEBIN_MAX_AGE * 2);
-
-        let mut server_txn = server.write(time_p1).await;
-        let admin = server_txn.internal_search_uuid(UUID_ADMIN).expect("failed");
-
-        let filt_i_rc = filter_all!(f_eq("class", PartialValue::new_class("recycled")));
-
-        let filt_i_ts = filter_all!(f_eq("class", PartialValue::new_class("tombstone")));
-
-        let filt_i_per = filter_all!(f_eq("class", PartialValue::new_class("person")));
-
-        // Create fake external requests. Probably from admin later
-        let me_rc = unsafe {
-            ModifyEvent::new_impersonate_entry(
-                admin.clone(),
-                filt_i_rc.clone(),
-                ModifyList::new_list(vec![Modify::Present(
-                    AttrString::from("class"),
-                    Value::new_class("recycled"),
-                )]),
-            )
-        };
-
-        let de_rc = unsafe { DeleteEvent::new_impersonate_entry(admin.clone(), filt_i_rc.clone()) };
-
-        let se_rc =
-            unsafe { SearchEvent::new_ext_impersonate_entry(admin.clone(), filt_i_rc.clone()) };
-
-        let sre_rc =
-            unsafe { SearchEvent::new_rec_impersonate_entry(admin.clone(), filt_i_rc.clone()) };
-
-        let rre_rc = unsafe {
-            ReviveRecycledEvent::new_impersonate_entry(
-                admin,
-                filter_all!(f_eq("name", PartialValue::new_iname("testperson1"))),
-            )
-        };
-
-        // Create some recycled objects
-        let e1 = entry_init!(
-            ("class", Value::new_class("object")),
-            ("class", Value::new_class("person")),
-            ("name", Value::new_iname("testperson1")),
-            (
-                "uuid",
-                Value::Uuid(uuid!("cc8e95b4-c24f-4d68-ba54-8bed76f63930"))
-            ),
-            ("description", Value::new_utf8s("testperson1")),
-            ("displayname", Value::new_utf8s("testperson1"))
-        );
-
-        let e2 = entry_init!(
-            ("class", Value::new_class("object")),
-            ("class", Value::new_class("person")),
-            ("name", Value::new_iname("testperson2")),
-            (
-                "uuid",
-                Value::Uuid(uuid!("cc8e95b4-c24f-4d68-ba54-8bed76f63932"))
-            ),
-            ("description", Value::new_utf8s("testperson2")),
-            ("displayname", Value::new_utf8s("testperson2"))
-        );
-
-        let ce = CreateEvent::new_internal(vec![e1, e2]);
-        let cr = server_txn.create(&ce);
-        assert!(cr.is_ok());
-
-        // Now we immediately delete these to force them to the correct state.
-        let de_sin = unsafe {
-            DeleteEvent::new_internal_invalid(filter!(f_or!([
-                f_eq("name", PartialValue::new_iname("testperson1")),
-                f_eq("name", PartialValue::new_iname("testperson2")),
-            ])))
-        };
-        assert!(server_txn.delete(&de_sin).is_ok());
-
-        // Can it be seen (external search)
-        let r1 = server_txn.search(&se_rc).expect("search failed");
-        assert!(r1.is_empty());
-
-        // Can it be deleted (external delete)
-        // Should be err-no candidates.
-        assert!(server_txn.delete(&de_rc).is_err());
-
-        // Can it be modified? (external modify)
-        // Should be err-no candidates
-        assert!(server_txn.modify(&me_rc).is_err());
-
-        // Can in be seen by special search? (external recycle search)
-        let r2 = server_txn.search(&sre_rc).expect("search failed");
-        assert!(r2.len() == 2);
-
-        // Can it be seen (internal search)
-        // Internal search should see it.
-        let r2 = server_txn
-            .internal_search(filt_i_rc.clone())
-            .expect("internal search failed");
-        assert!(r2.len() == 2);
-
-        // There are now two paths forward
-        //  revival or purge!
-        assert!(server_txn.revive_recycled(&rre_rc).is_ok());
-
-        // Not enough time has passed, won't have an effect for purge to TS
-        assert!(server_txn.purge_recycled().is_ok());
-        let r3 = server_txn
-            .internal_search(filt_i_rc.clone())
-            .expect("internal search failed");
-        assert!(r3.len() == 1);
-
-        // Commit
-        assert!(server_txn.commit().is_ok());
-
-        // Now, establish enough time for the recycled items to be purged.
-        let server_txn = server.write(time_p2).await;
-
-        //  purge to tombstone, now that time has passed.
-        assert!(server_txn.purge_recycled().is_ok());
-
-        // Should be no recycled objects.
-        let r4 = server_txn
-            .internal_search(filt_i_rc.clone())
-            .expect("internal search failed");
-        assert!(r4.is_empty());
-
-        // There should be one tombstone
-        let r5 = server_txn
-            .internal_search(filt_i_ts.clone())
-            .expect("internal search failed");
-        assert!(r5.len() == 1);
-
-        // There should be one entry
-        let r6 = server_txn
-            .internal_search(filt_i_per.clone())
-            .expect("internal search failed");
-        assert!(r6.len() == 1);
-
-        assert!(server_txn.commit().is_ok());
-    }
-
-    // The delete test above should be unaffected by recycle anyway
-    #[qs_test]
-    async fn test_qs_recycle_advanced(server: &QueryServer) {
-        // Create items
-        let mut server_txn = server.write(duration_from_epoch_now()).await;
-        let admin = server_txn.internal_search_uuid(UUID_ADMIN).expect("failed");
-
-        let e1 = entry_init!(
-            ("class", Value::new_class("object")),
-            ("class", Value::new_class("person")),
-            ("name", Value::new_iname("testperson1")),
-            (
-                "uuid",
-                Value::Uuid(uuid!("cc8e95b4-c24f-4d68-ba54-8bed76f63930"))
-            ),
-            ("description", Value::new_utf8s("testperson1")),
-            ("displayname", Value::new_utf8s("testperson1"))
-        );
-        let ce = CreateEvent::new_internal(vec![e1]);
-
-        let cr = server_txn.create(&ce);
-        assert!(cr.is_ok());
-        // Delete and ensure they became recycled.
-        let de_sin = unsafe {
-            DeleteEvent::new_internal_invalid(filter!(f_eq(
-                "name",
-                PartialValue::new_iname("testperson1")
-            )))
-        };
-        assert!(server_txn.delete(&de_sin).is_ok());
-        // Can in be seen by special search? (external recycle search)
-        let filt_rc = filter_all!(f_eq("class", PartialValue::new_class("recycled")));
-        let sre_rc = unsafe { SearchEvent::new_rec_impersonate_entry(admin, filt_rc.clone()) };
-        let r2 = server_txn.search(&sre_rc).expect("search failed");
-        assert!(r2.len() == 1);
-
-        // Create dup uuid (rej)
-        // After a delete -> recycle, create duplicate name etc.
-        let cr = server_txn.create(&ce);
-        assert!(cr.is_err());
-
-        assert!(server_txn.commit().is_ok());
-    }
-
-    #[qs_test]
     async fn test_name_to_uuid(server: &QueryServer) {
         let mut server_txn = server.write(duration_from_epoch_now()).await;
 
@@ -3363,79 +2556,6 @@ mod tests {
         // Uuid is not syntax normalised (but exists)
         let r4 = server_txn.uuid_to_rdn(uuid!("CC8E95B4-C24F-4D68-BA54-8BED76F63930"));
         assert!(r4.unwrap() == "spn=testperson1@example.com");
-    }
-
-    #[qs_test]
-    async fn test_uuid_to_star_recycle(server: &QueryServer) {
-        let mut server_txn = server.write(duration_from_epoch_now()).await;
-
-        let e1 = entry_init!(
-            ("class", Value::new_class("object")),
-            ("class", Value::new_class("person")),
-            ("class", Value::new_class("account")),
-            ("name", Value::new_iname("testperson1")),
-            (
-                "uuid",
-                Value::Uuid(uuid!("cc8e95b4-c24f-4d68-ba54-8bed76f63930"))
-            ),
-            ("description", Value::new_utf8s("testperson1")),
-            ("displayname", Value::new_utf8s("testperson1"))
-        );
-
-        let tuuid = uuid!("cc8e95b4-c24f-4d68-ba54-8bed76f63930");
-
-        let ce = CreateEvent::new_internal(vec![e1]);
-        let cr = server_txn.create(&ce);
-        assert!(cr.is_ok());
-
-        assert!(server_txn.uuid_to_rdn(tuuid) == Ok("spn=testperson1@example.com".to_string()));
-
-        assert!(
-            server_txn.uuid_to_spn(tuuid)
-                == Ok(Some(Value::new_spn_str("testperson1", "example.com")))
-        );
-
-        assert!(server_txn.name_to_uuid("testperson1") == Ok(tuuid));
-
-        // delete
-        let de_sin = unsafe {
-            DeleteEvent::new_internal_invalid(filter!(f_eq(
-                "name",
-                PartialValue::new_iname("testperson1")
-            )))
-        };
-        assert!(server_txn.delete(&de_sin).is_ok());
-
-        // all should fail
-        assert!(
-            server_txn.uuid_to_rdn(tuuid)
-                == Ok("uuid=cc8e95b4-c24f-4d68-ba54-8bed76f63930".to_string())
-        );
-
-        assert!(server_txn.uuid_to_spn(tuuid) == Ok(None));
-
-        assert!(server_txn.name_to_uuid("testperson1").is_err());
-
-        // revive
-        let admin = server_txn.internal_search_uuid(UUID_ADMIN).expect("failed");
-        let rre_rc = unsafe {
-            ReviveRecycledEvent::new_impersonate_entry(
-                admin,
-                filter_all!(f_eq("name", PartialValue::new_iname("testperson1"))),
-            )
-        };
-        assert!(server_txn.revive_recycled(&rre_rc).is_ok());
-
-        // all checks pass
-
-        assert!(server_txn.uuid_to_rdn(tuuid) == Ok("spn=testperson1@example.com".to_string()));
-
-        assert!(
-            server_txn.uuid_to_spn(tuuid)
-                == Ok(Some(Value::new_spn_str("testperson1", "example.com")))
-        );
-
-        assert!(server_txn.name_to_uuid("testperson1") == Ok(tuuid));
     }
 
     #[qs_test]
